requirements.txt
amqp==5.0.6
astroid==2.8.5
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Flask==2.0.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mpmath==1.2.1
numpy==1.21.4
pandas==1.3.4
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
pylint==2.11.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sympy==1.9
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
